squash

+++ Informix.DB2.LinqService/Tests/Linq/TypesTests/Tests.Linq.TypesTests.UpdateBinary2(Informix.DB2.LinqService).sql
@@ -8,7 +8,7 @@ SET     @p = 1
 UPDATE
 	LinqDataTypes t
 SET
-	BinaryValue = @BinaryValue::BYTE
+	BinaryValue = @BinaryValue
 WHERE
 	t.ID = @p
 
@@ -22,7 +22,7 @@ SET     @p = 2
 UPDATE
 	LinqDataTypes t
 SET
-	BinaryValue = @BinaryValue::BYTE
+	BinaryValue = @BinaryValue
 WHERE
 	t.ID = @p
