+++ Informix.DB2.LinqService/Tests/UserTests/Issue421/Tests.UserTests.Issue421.Test2(Informix.DB2.LinqService).sql
@@ -27,7 +27,7 @@ INSERT INTO BlobClass
 VALUES
 (
 	1,
-	@BlobValue::byte
+	@BlobValue
 )
 
 BeforeExecute
@@ -49,7 +49,7 @@ SET     @BlobValue = {3,2,1}
 UPDATE
 	BlobClass t1
 SET
-	BlobValue = @BlobValue::byte
+	BlobValue = @BlobValue
 WHERE
 	t1.Id = 1
