+++ Informix.DB2.LinqService/Tests/xUpdate/InsertTests/Tests.xUpdate.InsertTests.Insert1(Informix.DB2.LinqService).sql
@@ -19,7 +19,7 @@ INSERT INTO Child
 VALUES
 (
 	1,
-	@id::Int
+	@id
 )
 
 BeforeExecute
