+++ Informix.DB2.LinqService/Tests/xUpdate/InsertTests/Tests.xUpdate.InsertTests.InsertArray2(Informix.DB2.LinqService).sql
@@ -21,7 +21,7 @@ VALUES
 (
 	1001,
 	't',
-	@arr::BYTE
+	@arr
 )
 
 BeforeExecute
