+++ Informix.DB2/Tests/Linq/TypesTests/Tests.Linq.TypesTests.UpdateBinary1(Informix.DB2).sql
@@ -6,7 +6,7 @@ SET     @BinaryValue = {1,2,3,4,5}
 UPDATE
 	LinqDataTypes t
 SET
-	BinaryValue = @BinaryValue::BYTE
+	BinaryValue = @BinaryValue
 WHERE
 	t.ID = 1
