+++ Informix.DB2/Tests/xUpdate/DeleteTests/Tests.xUpdate.DeleteTests.Delete5(Informix.DB2).sql
@@ -26,7 +26,7 @@ INSERT INTO Parent
 )
 VALUES
 (
-	@ParentID::Int,
+	@ParentID,
 	1
 )
 
@@ -42,7 +42,7 @@ INSERT INTO Parent
 )
 VALUES
 (
-	@ParentID::Int,
+	@ParentID,
 	1
 )
